requirements.txt
rich
requests
loguru
einops
pillow
gradio